refactor monitor for capnp or other serialized input
remove json wording, header now generic serialized data.
